Add validation to production build script
Validation has been added to the build script when the "prod" target is
selected. We now ensure that all expected environment variables are
set, and that no extra environment variables are present (which might
indicate that the wrong configuration file is being used).

A new "dist" target has been added to preserve the ability to build a
"production-like" build without this validation.

Closes #15003